feat: widen block-id storage to sixteen bits (#25)
The kernel moved its block ids to sixteen bits some time ago; this
package still stored them one byte per block, in both its in-memory
generation buffer and its persisted chunk format. Nothing truncated at
today's registry size, but past two hundred and fifty-six entries ids
would wrap silently, modulo two hundred and fifty-six, with no error
raised in either the generation or the save path — corrupt terrain and
corrupt saved worlds, produced by adding content, which is exactly
what feature parity means doing.

The buffer and every generation pass now hold sixteen-bit elements.
Light grids and pixel buffers are deliberately untouched: those really
are one byte per voxel and have nothing to do with block ids.

The persisted format is version two, two bytes per element, packed
through an explicit view rather than by aliasing a typed array over
the buffer — the latter reads host-endian, which the kernel's own
encoder deliberately avoids, so this matches it. The element width is
taken from the kernel's exported constant rather than written as a
second literal two, so the two cannot drift apart.

Version one survives read-only, with a migration applied on load. The
save library refuses any version but the current one and has no
automatic upgrade path, so a consumer that wants old worlds to keep
opening has to do this itself: try the current schema, and on a
version mismatch decode under the old one and widen. Every write is
version two, so worlds upgrade the next time they are saved rather
than being rewritten under a bare load. Breaking existing saves was
the one outcome worse than the bug.

A test round-trips two ids that byte storage cannot hold: three hundred,
which could only ever have been stored as forty-four, and the sixteen-bit
ceiling itself. Another seals
a hand-authored version-one envelope into storage and proves it loads,
migrates, and is rewritten as version two.

The golden digests move because the hashed byte sequence doubled in
width, not because terrain changed: every other golden invariant —
bedrock floor, sea level, water, ore and tree counts — is unchanged.
Their hashing also now uses the same explicit packing as the save
format, rather than reading the buffer's host-endian memory.

+++ apps/preview-terrain/stats.ts
@@ -473,7 +473,7 @@ const localReport = (params: WorldParams, options: StatsOptions): ReadonlyArray<
  * buffer, which is the chunk manager's job), so the real world will be at least
  * this bad and probably worse.
  */
-const largestLeafSheet = (blocks: Uint8Array): number => {
+const largestLeafSheet = (blocks: Uint16Array): number => {
   let largest = 0
   const seen = new Uint8Array(CHUNK_SIZE_XZ * CHUNK_SIZE_XZ)
   const stack: Array<number> = []
